Flags in the mesh were reversed. Corrected that.
Added an object to define the arrayfire mesh called AfMesh2D. This object is used most frequently for simulations. As is this class accepts a Mesh2D object and converts the mesh to arrayfire. Later, this class should be changed to inherit Mesh2D and define it internally.

The a new line to expose AfMesh2D was added to __init__.py together with the "AfMesh2D" string to the __all__ list.