Added the field.py file containing a series of classes to define a 2D field. Each class has its properties and functionality,
The class StoreField contains the methods for Field2D to store the 2D field in a numpy ndarray.

The class Field2D initiates the storage for the 2D numpy ndarray. Through StoreField we can store the field.

The class AfField2D adds the possibility to convert the field between numpy ndarray and arrayfire Array. This feature is more commonly exploited for simulations using arrayfire.

AfLandscapeField2D inherits AfField2D and adds functions to interpret the modulation configuration object. This allows the plotting of the background state and the landscape modulation of the field.

The class AfCoupled2D extends the functionality of the AfLandscapeField2D to a 2 field configuration. Later this class should accept the first field by means of inheritance.

The class CoupledPlotting introduces plotting functions to any Coupled Field object.

Lastly, the NotebookAfCoupledSimulation2D inherits both CoupledPlotting and AfCoupled2D to have a single object with plotting and generation of arrayfire landscaped 2D fields.

To conclude, in the __init__.py we exposed the functions Field2D, AfField2D, AfCoupled2D and NotebookAfCoupledSimulation2D.