The class LatticeConfig carries the basic properties of a lattice such as lattice parameters, amplitude constant and rotation.
The class Lattice inherits LatticeConfig and introduces the lattice_method function to plot the landscape over a Mesh2D. Later should be adapted to use @dataclass.

The class GaussianBeamConfig stores the gaussian config and carries the gaussian modulation function.

The class LatticeGaussianBeamConfig inherits both GaussianBeamConfig and Lattice and introduces a function to store the properties of both the gaussian profile and lattice landscape.

The class LatticeGaussianCoupledConfig extends the functionality of LatticeGaussianBeamConfig to two beams. Later this method should define the first beam using inheritance and introduce the second beam only.

To conclude, the __init__.py file exposes the classes LatticeConfig, GaussianBeamConfig, LatticeGaussianBeamConfig and LatticeGaussianCoupledConfig.