The order of the input parameters of the decorator check_field were exchanged (field <-> filename) and with them every function calling the decorator.
The Field2D now has a method to copy the field attribute onto an input_field to be called only at the start of a simulation in order to preserve initial states.

The class AfCoupled2D now has a method to call copy_input_field of both fields.

The 2D plotting functions were passed onto a class Plotting2D which is inherited by CoupledPlotting.

The 3D plotting functions are defined in Plotting3D which is ingerited by CoupledPlotting.

Inside CoupledPlotting and by calling the inherited classes, we have a function called plot_fields which plots a 2x2 grid with the intensity field and phase profiles of the fields. A second function with the name plot_input_output_3d is defined to plot the input intensity field in 2d and the output intensity field in 2d and 3d.