Separated the field modules into distinct .py files:
The field carries the field and coupled field method. This should be further segmented into a field containing the single field method and a coupled fields containing more than one field methods.

The field plotting methods were moved to the field_plotting.py file while the junction of the plotting methods and fields are found in the notebook_field.py. This separation is given by the requirements of running simulations and code without plotting functionality and plotting required in notebook form.