Commented the class and internal functions.
Removed a decorator wrapping the basic single ax plotting of the intensity field and phase profile.

The basic functions work by changing the figure and axis inplace. As such, the return of figure and axis were removed together with the option to show the plot. Future additions could benefit from a function initializing a figure and calling these methods to plot the fields. Similar changes were made to the 3d plotting method.

The 3d plotting method now has a parameter zlim to define the limit of the height of the plotting.

Inside CoupledPlotting the function plot_input_output_3d is now called plot_IO_3d.